[#64] SwiftLint 구조 경고 해결 및 iPhone Portrait 전용 설정
