+++ Codive/Features/Closet/Data/ClothAPIService.swift
@@ -84,8 +84,8 @@ struct ClothUpdateAPIRequest {
 
 final class ClothAPIService: ClothAPIServiceProtocol {
 
-    private let client: Client
-    private let jsonDecoder: JSONDecoder
+    let client: Client
+    let jsonDecoder: JSONDecoder
 
     init(tokenProvider: TokenProvider = KeychainTokenProvider()) {
         self.client = CodiveAPIProvider.createClient(
@@ -355,89 +355,6 @@ extension ClothAPIService {
             
         case .undocumented(statusCode: let code, _):
             throw LookBookAPIError.serverError(statusCode: code, message: "룩북 목록 조회 실패")
-        }
-    }
-}
-
-// MARK: - Private Helpers
-
-private extension ClothAPIService {
-
-    func calculateMD5(from data: Data) -> String {
-        let digest = Insecure.MD5.hash(data: data)
-        return Data(digest).base64EncodedString()
-    }
-
-    func extractFinalUrl(from presignedUrl: String) -> String {
-        guard let url = URL(string: presignedUrl),
-              var components = URLComponents(url: url, resolvingAgainstBaseURL: false) else {
-            return presignedUrl
-        }
-        components.query = nil
-        return components.string ?? presignedUrl
-    }
-
-    func mapSeasonToCreatePayload(_ season: Season) -> Components.Schemas.ClothCreateRequest.seasonsPayloadPayload {
-        switch season {
-        case .spring: return .SPRING
-        case .summer: return .SUMMER
-        case .fall: return .FALL
-        case .winter: return .WINTER
-        }
-    }
-
-    func mapSeasonToUpdatePayload(_ season: Season) -> Components.Schemas.ClothUpdateRequest.seasonsPayloadPayload {
-        switch season {
-        case .spring: return .SPRING
-        case .summer: return .SUMMER
-        case .fall: return .FALL
-        case .winter: return .WINTER
-        }
-    }
-
-    func mapSeasonToQueryParam(_ season: Season) -> Operations.Cloth_getClothes.Input.Query.seasonsPayloadPayload {
-        switch season {
-        case .spring: return .SPRING
-        case .summer: return .SUMMER
-        case .fall: return .FALL
-        case .winter: return .WINTER
-        }
-    }
-
-    func mapSeasonToSearchQueryParam(_ season: Season) -> Operations.Search_searchClothes.Input.Query.seasonsPayloadPayload {
-        switch season {
-        case .spring: return .SPRING
-        case .summer: return .SUMMER
-        case .fall: return .FALL
-        case .winter: return .WINTER
-        }
-    }
-}
-
-// MARK: - ClothAPIError
-
-enum ClothAPIError: LocalizedError {
-    case presignedUrlMismatch
-    case invalidUrl
-    case invalidResponse
-    case s3UploadFailed(statusCode: Int)
-    case noClothIdsReturned
-    case serverError(statusCode: Int, message: String)
-
-    var errorDescription: String? {
-        switch self {
-        case .presignedUrlMismatch:
-            return "Presigned URL 개수가 요청한 이미지 개수와 일치하지 않습니다."
-        case .invalidUrl:
-            return "유효하지 않은 URL입니다."
-        case .invalidResponse:
-            return "서버 응답을 처리할 수 없습니다."
-        case .s3UploadFailed(let statusCode):
-            return "S3 업로드 실패 (상태 코드: \(statusCode))"
-        case .noClothIdsReturned:
-            return "서버에서 생성된 옷 ID를 반환하지 않았습니다."
-        case .serverError(let statusCode, let message):
-            return "서버 오류 (\(statusCode)): \(message)"
         }
     }
 }
