+++ Codive/Features/Profile/MyProfile/Data/ProfileAPIService.swift
@@ -32,8 +32,8 @@ protocol ProfileAPIServiceProtocol {
 // MARK: - Profile API Service Implementation
 
 final class ProfileAPIService: ProfileAPIServiceProtocol {
-    private let client: Client
-    private let jsonDecoder: JSONDecoder
+    let client: Client
+    let jsonDecoder: JSONDecoder
 
     init(tokenProvider: TokenProvider = KeychainTokenProvider()) {
         self.client = CodiveAPIProvider.createClient(
@@ -290,7 +290,7 @@ final class ProfileAPIService: ProfileAPIServiceProtocol {
         }
     }
 
-    private func uploadImageToS3(presignedUrl: String, imageData: Data, contentMD5: String) async throws {
+    func uploadImageToS3(presignedUrl: String, imageData: Data, contentMD5: String) async throws {
         guard let url = URL(string: presignedUrl) else {
             throw ProfileAPIError.invalidUrl
         }
@@ -309,149 +309,6 @@ final class ProfileAPIService: ProfileAPIServiceProtocol {
 
         guard (200...299).contains(httpResponse.statusCode) else {
             throw ProfileAPIError.s3UploadFailed(statusCode: httpResponse.statusCode)
-        }
-    }
-}
-
-extension ProfileAPIService {
-    /// 나의 최애 코디 조회
-    func fetchMyFavoriteCoordinate() async throws -> [MyFavoriteLookBookResponseDTO] {
-        let input = Operations.Coordinate_getFavoriteCoordinates.Input()
-        let response = try await client.Coordinate_getFavoriteCoordinates(input)
-        
-        switch response {
-        case .ok(let okResponse):
-            let data = try await Data(collecting: okResponse.body.any, upTo: .max)
-            let jsonDecoder = JSONDecoderFactory.makeAPIDecoder()
-            
-            let decoded = try jsonDecoder.decode(Components.Schemas.BaseResponseListFavoriteCoordinateResponse.self, from: data)
-            
-            let items = decoded.result ?? []
-            
-            return items.map { item in
-                MyFavoriteLookBookResponseDTO(
-                    coordinateId: item.coordinateId ?? 0,
-                    imageUrl: item.imageUrl ?? "",
-                    coordinateName: item.coordinateName ?? ""
-                )
-            }
-        case .undocumented(statusCode: let code, _):
-            throw ProfileAPIError.serverError(statusCode: code, message: "최애 코디 조회 실패 (상태코드: \(code))")
-        }
-    }
-    
-    func fetchCoordinatePreview(coordinateId: Int64) async throws -> CoordinatePreviewResponseDTO {
-        let input = Operations.Coordinate_getCoordinatePreview.Input(
-            path: .init(
-                coordinateId: coordinateId
-            )
-        )
-        
-        let response = try await client.Coordinate_getCoordinatePreview(input)
-        
-        switch response {
-        case .ok(let okResponse):
-            let data = try await Data(collecting: okResponse.body.any, upTo: .max)
-            
-            let decoded = try jsonDecoder.decode(Components.Schemas.BaseResponseCoordinatePreviewResponse.self, from: data)
-            
-            guard let item = decoded.result else {
-                throw LookBookAPIError.invalidResponse
-            }
-            
-            return CoordinatePreviewResponseDTO(
-                coordinateId: item.coordinateId ?? 0,
-                imageUrl: item.imageUrl ?? "",
-                coordinateName: item.coordinateName ?? "",
-                coordinateMemo: item.coordinateMemo ?? ""
-            )
-            
-        case .undocumented(statusCode: let code, _):
-            throw ProfileAPIError.serverError(statusCode: code, message: "코디 preview 조회 실패")
-        }
-    }
-    
-    func fetchCoordinateDetail(
-        coordinateId: Int64
-    ) async throws -> [CoordinateDetailResponseDTO] {
-        let input = Operations.Coordinate_getCoordinateDetails.Input(
-            path: .init(
-                coordinateId: coordinateId
-            )
-        )
-        
-        let response = try await client.Coordinate_getCoordinateDetails(input)
-        
-        switch response {
-        case .ok(let okResponse):
-            let data = try await Data(collecting: okResponse.body.any, upTo: .max)
-            
-            let decoded = try jsonDecoder.decode(Components.Schemas.BaseResponseListCoordinateDetailsListResponse.self, from: data)
-            
-            let items = decoded.result ?? []
-            
-            return items.map { item in
-                CoordinateDetailResponseDTO(
-                    coordinateClothId: item.coordinateClothId ?? 0,
-                    locationX: item.locationX ?? 0,
-                    locationY: item.locationY ?? 0,
-                    ratio: item.ratio ?? 1.0,
-                    degree: item.degree ?? 0,
-                    order: item.order ?? 0,
-                    clothId: item.clothId ?? 0,
-                    imageUrl: item.imageUrl ?? "",
-                    brand: item.brand ?? "",
-                    name: item.name ?? "",
-                    category: item.category ?? "",
-                    parentCategory: item.parentCategory ?? ""
-                )
-            }
-            
-        case .undocumented(statusCode: let code, _):
-            throw ProfileAPIError.serverError(statusCode: code, message: "코디 detail 조회 실패")
-        }
-    }
-}
-
-// MARK: - Private Helpers
-
-private extension ProfileAPIService {
-    func calculateMD5(from data: Data) -> String {
-        let digest = Insecure.MD5.hash(data: data)
-        return Data(digest).base64EncodedString()
-    }
-
-    func extractFinalUrl(from presignedUrl: String) -> String {
-        guard let url = URL(string: presignedUrl),
-              var components = URLComponents(url: url, resolvingAgainstBaseURL: false) else {
-            return presignedUrl
-        }
-        components.query = nil
-        return components.string ?? presignedUrl
-    }
-}
-
-// MARK: - Profile API Error
-
-enum ProfileAPIError: LocalizedError {
-    case serverError(statusCode: Int, message: String)
-    case invalidResponse
-    case invalidUrl
-    case s3UploadFailed(statusCode: Int)
-    case noData
-
-    var errorDescription: String? {
-        switch self {
-        case .serverError(let statusCode, let message):
-            return "서버 오류 (\(statusCode)): \(message)"
-        case .invalidResponse:
-            return "올바르지 않은 응답 형식입니다"
-        case .invalidUrl:
-            return "유효하지 않은 URL입니다"
-        case .s3UploadFailed(let statusCode):
-            return "S3 업로드 실패 (상태 코드: \(statusCode))"
-        case .noData:
-            return "응답 데이터가 없습니다"
         }
     }
 }
